Description: dynamic pull daudio_host process error
Match-id-02d9514851415a55b617bab208ddbdd96ad139e0

+++ hdf_service/distributed_audio/hdi_service/common/utils/src/daudio_utils.cpp
@@ -47,7 +47,7 @@ std::string GetAnonyString(const std::string &value)
 
 int32_t GetAudioParamStr(const std::string &params, const std::string &key, std::string &value)
 {
-    int32_t step = key.size();
+    size_t step = key.size();
     if (step >= params.size()) {
         return ERR_DH_AUDIO_HDF_FAIL;
     }
